+++ src/main/kotlin/info/dropy/dropy/authentication/Customer/data/OrderItemDto.kt
@@ -3,7 +3,7 @@ package info.dropy.dropy.authentication.Customer.data
 
 data class OrderItemDto (
     val quantity: Int,
-    val product: Int,
-    val shop: Int,
-    val customer: Int
+    val product: Long,
+    val shop: Long,
+    val customer: Long
         )
